BOCM-80 - moves date formatting function to own file for reusability
